pyproject.toml
[project]
name = "openai-thread-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-core>=1.39.0",
    "azure-identity>=1.25.3",
    "lagom>=2.7.7",
    "openai>=2.32.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "openai_thread_id/protocols/*.py",
    "openai_thread_id/hosting.py"
]

